backend/itenrary_agent/requirements.txt

fastapi
google-adk
google-cloud-aiplatform
google-genai
fastapi[standard]
uvicorn
pandas
